requirements.txt
aiosmtplib==1.0.6
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
dnspython==1.16.0
email-validator==1.0.4
fastapi==0.33.0
h11==0.8.1
idna==2.8
passlib==1.7.1
pycparser==2.19
pydantic==0.30
PyJWT==1.7.1
pymongo==3.8.0
python-multipart==0.0.5
six==1.12.0
starlette==0.12.0
uvicorn==0.8.3
websockets==7.0
